requirements.txt
flask
gunicorn
python-telegram-bot>=20.0
requests
schedule
aiohttp
psycopg2-binary
